requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
gunicorn==21.2.0
stripe==7.8.0
python-dotenv==1.0.0
openai==1.6.1
mysql-connector-python==8.2.0
pymysql==1.1.0
redis==5.0.1
httpx==0.27.0
WTForms==3.1.1
werkzeug==3.0.1
cryptography==41.0.7
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
locust==2.20.0
pytest-cov==4.1.0
